Split backend-availability helpers into their own support module
Reduce the responsibility count of the tee-profile concurrency scaffolding
by moving the backend-availability and Hypothesis parametrization helpers
out of the lock/selector harness.

- Add `cuprum/unittests/_tee_profile_backend_support.py` holding
  `_rust_backend_available`, `_available_backend_names`, `_alternate_backend`,
  `_backend_pairs`, `_backend_lists`, and `_backends_strategy` (a
  self-contained cluster that only referenced each other).
- Drop those helpers, and the now-unused `pytest`, `hypothesis.strategies`,
  and `ParameterSet` imports, from `_tee_profile_concurrency_support.py`,
  which retains the instrumented lock, selectors, race harness, and the
  thread join/assert helper.
- Point the concurrency and reentrancy test modules at the new module,
  preserving all backend-availability and Hypothesis parametrization
  behaviour.
- Regenerate the wheel-layout snapshot and update the developers' guide to
  name both support modules.
