+++ cuprum/unittests/test_tee_profile_worker_selector_reentrancy.py
@@ -17,7 +17,7 @@ import pytest
 from hypothesis import HealthCheck, given, settings
 
 from benchmarks import tee_profile_worker
-from cuprum.unittests._tee_profile_concurrency_support import _backends_strategy
+from cuprum.unittests._tee_profile_backend_support import _backends_strategy
 
 if typ.TYPE_CHECKING:
     from syrupy.assertion import SnapshotAssertion
